pyproject.toml
[project]
name = "text-visual"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "hanlp>=2.1.1",
]

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "pip>=25.2",
    "pydantic-to-typescript>=2.0.0",
    "ruff>=0.12.9",
]

[tool.ruff]
line-length = 120
indent-width = 4

# Assume Python 3.12
target-version = "py312"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`) codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F", "I"]
